% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\alias{dep_corejs}
\alias{dep_semiotic}
\title{Dependencies for g2, bizcharts, and data-set}
\usage{
dep_corejs()

dep_semiotic()
}
\value{
\code{htmltools::htmlDependency}
}
\description{
Dependencies for g2, bizcharts, and data-set
}
