% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{rem_lines}
\alias{rem_lines}
\title{Title}
\usage{
rem_lines(remwidget = NULL, lineDataAccessor = NULL, lineType = NULL,
  lineStyle = NULL, lineClass = NULL, lineIDAccessor = NULL)
}
\arguments{
\item{lineIDAccessor}{}
}
\description{
Title
}
