% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frames.R
\name{XYFrame}
\alias{XYFrame}
\title{Title}
\usage{
XYFrame(data = NULL, title = NULL, xAccessor = NULL, xExtent = NULL,
  yAccessor = NULL, yExtent = NULL, width = NULL, height = NULL,
  margin = NULL, elementId = NULL)
}
\arguments{
\item{yExent}{}
}
\description{
Title
}
